refactor(prefill): collapse layer-completion into a model-local _layer_completion module
The whole layer-completion stack is used ONLY by the deepseek_v3_d_p pipelined-prefill
runner (no other ttnn.layer_completion / internal/disaggregation consumers). It was spread
across tt_metal core, the ttnn public API, and models/demos/test. Collapse it to a single
model-local feature (mirrors #48300's containment):

- All C++ (message, ring_layout, queue, reorder_buffer, router, test-only consumer) moves to
  models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/, bound by one
  NB_MODULE(_layer_completion), exposed via the package __init__.py.
- Built as a standalone nanobind .so via the ttnn CMake hook (only scope with nanobind +
  TT::Metalium); links only TT::Metalium, emitted to build tree + symlinked next to the runner.
  NOT in _ttnn, NOT on the ttnn API.
- Removed from tt_metal (api/internal/disaggregation + tt_metal/disaggregation + distributed
  sources/include-hack), ttnn (binding + registration + __init__), and reverted the
  validate_includes.py 'internal' allowlist.
- Runner imports lazily from the package; C++ unit + MPI routing tests compile the feature
  impls and include from the feature dir.

Verified: full build clean, _ttnn.so no longer contains the binding, ttnn.layer_completion
gone, the package imports all three classes, host wiring test passes.
