+++ models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/layer_completion_queue.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <internal/disaggregation/layer_completion_queue.hpp>
+#include "layer_completion_queue.hpp"
 
 #include <chrono>
 #include <cstring>
