+++ models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/layer_completion_reorder_buffer.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <internal/disaggregation/layer_completion_reorder_buffer.hpp>
+#include "layer_completion_reorder_buffer.hpp"
 
 #include <tt-logger/tt-logger.hpp>
