+++ models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/layer_completion_reorder_buffer.hpp
@@ -24,7 +24,7 @@
 #include <map>
 #include <vector>
 
-#include <internal/disaggregation/layer_completion_message.hpp>
+#include "layer_completion_message.hpp"
 
 namespace tt::tt_metal::distributed {
