+++ models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/layer_completion_ring_layout.hpp
@@ -20,7 +20,7 @@
 #include <cstddef>
 #include <cstdint>
 
-#include <internal/disaggregation/layer_completion_message.hpp>
+#include "layer_completion_message.hpp"
 
 namespace tt::tt_metal::distributed {
