+++ models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill/layer_completion_router.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <internal/disaggregation/layer_completion_router.hpp>
+#include "layer_completion_router.hpp"
 
 #include <array>
 #include <chrono>
@@ -15,9 +15,9 @@
 #include <tt-metalium/distributed_context.hpp>
 
 #include <internal/service/inter_process_counter_channel.hpp>
-#include <internal/disaggregation/layer_completion_message.hpp>
-#include <internal/disaggregation/layer_completion_queue.hpp>
-#include <internal/disaggregation/layer_completion_reorder_buffer.hpp>
+#include "layer_completion_message.hpp"
+#include "layer_completion_queue.hpp"
+#include "layer_completion_reorder_buffer.hpp"
 
 namespace tt::tt_metal::distributed {
