+++ tests/tt_metal/distributed/test_layer_completion_queue_mpsc.cpp
@@ -10,8 +10,8 @@
 #include <unordered_set>
 #include <vector>
 
-#include <internal/disaggregation/layer_completion_message.hpp>
-#include <internal/disaggregation/layer_completion_queue.hpp>
+#include <layer_completion_message.hpp>
+#include <layer_completion_queue.hpp>
 
 namespace tt::tt_metal::distributed::test {
