+++ tests/tt_metal/multihost/layer_completion/CMakeLists.txt
@@ -2,11 +2,17 @@ add_library(multi_host_layer_completion OBJECT)
 add_library(TT::Metalium::Test::Distributed::MultihostLayerCompletion ALIAS multi_host_layer_completion)
 TT_ENABLE_UNITY_BUILD(multi_host_layer_completion)
 
+set(LAYER_COMPLETION_DIR ${PROJECT_SOURCE_DIR}/models/demos/deepseek_v3_d_p/tt/runners/pipelined_prefill)
 target_sources(
     multi_host_layer_completion
     PRIVATE
         main.cpp
         test_layer_completion_routing.cpp
+        # layer-completion impls live with the pipelined-prefill feature; compile them in so this
+        # MPI routing test links without the feature's standalone .so.
+        ${LAYER_COMPLETION_DIR}/layer_completion_queue.cpp
+        ${LAYER_COMPLETION_DIR}/layer_completion_reorder_buffer.cpp
+        ${LAYER_COMPLETION_DIR}/layer_completion_router.cpp
 )
 
 target_include_directories(
@@ -15,6 +21,8 @@ target_include_directories(
         "$<TARGET_PROPERTY:Metalium::Metal,INCLUDE_DIRECTORIES>"
         ${PROJECT_SOURCE_DIR}/tests
         ${PROJECT_SOURCE_DIR}/tests/tt_metal/multihost
+        ${PROJECT_SOURCE_DIR}/tt_metal/distributed # named_shm.hpp (used by the queue impl)
+        ${LAYER_COMPLETION_DIR} # layer-completion feature headers
 )
 
 target_link_libraries(multi_host_layer_completion PRIVATE test_metal_common_libs)
