+++ tests/ttnn/unit_tests/base_functionality/test_prefill_layer_completion_wiring.py
@@ -12,6 +12,7 @@ import os
 import time
 
 import ttnn
+from models.demos.deepseek_v3_d_p.tt.runners.pipelined_prefill import LayerCompletionQueue, LayerCompletionRouter
 
 from models.demos.deepseek_v3_d_p.tt.runners.prefill_runner import build_layer_completion_sink
 
@@ -29,11 +30,11 @@ def test_sink_pushes_dense_seq_and_router_orders_them():
     _unlink(ring)
     _unlink(sched)
 
-    router = ttnn.layer_completion.LayerCompletionRouter(
+    router = LayerCompletionRouter(
         rank=0, world_size=1, master_rank=0, ring_shm_name=ring, scheduler_channel_shm_name=sched
     )
     consumer = ttnn.InterProcessCounterChannel.connect(sched, connect_timeout_ms=5000)
-    producer = ttnn.layer_completion.LayerCompletionQueue.connect(ring, connect_timeout_ms=5000)
+    producer = LayerCompletionQueue.connect(ring, connect_timeout_ms=5000)
 
     num_layers = 4
     sink = build_layer_completion_sink(producer, source_rank=0, num_layers=num_layers)
