+++ ttnn/cpp/ttnn-nanobind/__init__.cpp
@@ -20,7 +20,6 @@
 #include "ttnn-nanobind/hd_socket.hpp"
 #include "ttnn-nanobind/d2d_stream_service.hpp"
 #include "ttnn-nanobind/counter_channel.hpp"
-#include "ttnn-nanobind/layer_completion.hpp"
 #include "ttnn-nanobind/h2d_stream_service.hpp"
 #include "ttnn-nanobind/d2h_stream_service.hpp"
 #include "ttnn-nanobind/mesh_socket.hpp"
@@ -258,7 +257,6 @@ NB_MODULE(_ttnn, mod) {
     auto m_d2d_stream_service =
         mod.def_submodule("d2d_stream_service", "ttnn persistent device-to-device streaming service");
     auto m_counter_channel = mod.def_submodule("counter_channel", "ttnn cross-process producer-counter channel");
-    auto m_layer_completion = mod.def_submodule("layer_completion", "ttnn pipelined-prefill layer-completion routing");
     auto m_mesh_socket = mod.def_submodule("mesh_socket", "ttnn mesh socket");
     auto m_profiler = mod.def_submodule("profiler", "Submodule defining the profiler");
     auto m_reports = mod.def_submodule("reports", "ttnn reports");
@@ -287,7 +285,6 @@ NB_MODULE(_ttnn, mod) {
     ttnn::d2h_stream_service::py_module_types(m_d2h_stream_service);
     ttnn::d2d_stream_service::py_module_types(m_d2d_stream_service);
     ttnn::counter_channel::py_module_types(m_counter_channel);
-    ttnn::layer_completion::py_module_types(m_layer_completion);
     ttnn::mesh_socket::py_module_types(m_mesh_socket);
     ttnn::reports::py_module_types(m_reports);
     ttnn::program_descriptors::py_module_types(m_program_descriptors);
@@ -329,7 +326,6 @@ NB_MODULE(_ttnn, mod) {
     ttnn::d2h_stream_service::py_module(m_d2h_stream_service);
     ttnn::d2d_stream_service::py_module(m_d2d_stream_service);
     ttnn::counter_channel::py_module(m_counter_channel);
-    ttnn::layer_completion::py_module(m_layer_completion);
     ttnn::mesh_socket::py_module(m_mesh_socket);
     ttnn::profiler::py_module(m_profiler);
     ttnn::reports::py_module(m_reports);
