--- ttnn/cpp/ttnn-nanobind/layer_completion.hpp
@@ -1,15 +1,0 @@
-// SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "ttnn-nanobind/nanobind_fwd.hpp"
-
-namespace ttnn::layer_completion {
-namespace nb = nanobind;
-
-void py_module_types(nb::module_& mod);
-void py_module(nb::module_& mod);
-
-}  // namespace ttnn::layer_completion
