+++ ttnn/ttnn/__init__.py
@@ -223,7 +223,6 @@ from ttnn._ttnn.counter_channel import (
     InterProcessCounterChannel,
 )
 
-from ttnn._ttnn import layer_completion
 
 from ttnn.types import (
     TILE_SIZE,
